requirements.txt
beautifulsoup4==4.12.2
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pydantic==2.5.0
PyYAML==6.0.1
six==1.13.0
soupsieve==1.9.5
webcolors==1.10
Werkzeug==0.16.0
